--- src/bun.js/bindings/webcore/weak_handle.h
@@ -1,7 +1,0 @@
-#pragma once
-
-#include "config.h"
-
-namespace JSC {
-
-} // namespace JSC
